Factor out common APDU code and normalize

+++ src/apdu_baking.h
@@ -1,9 +1,10 @@
 #pragma once
 
+#include <stddef.h>
 #include <stdint.h>
 
-unsigned int handle_apdu_reset(uint8_t instruction);
-unsigned int handle_apdu_query_auth_key(uint8_t instruction);
-unsigned int handle_apdu_main_hwm(uint8_t instruction);
-unsigned int handle_apdu_all_hwm(uint8_t instruction);
-unsigned int handle_apdu_deauthorize(uint8_t instruction);
+size_t handle_apdu_reset(uint8_t instruction);
+size_t handle_apdu_query_auth_key(uint8_t instruction);
+size_t handle_apdu_main_hwm(uint8_t instruction);
+size_t handle_apdu_all_hwm(uint8_t instruction);
+size_t handle_apdu_deauthorize(uint8_t instruction);
